GeoFleX: Add auto calculation of minimum samples per simulation.
This is based on the sample size needed to safely evaluate the coverage. The user can either set the samples to "auto", in which case the minimum required samples is used. Otherwise they can set it to a value and if the value is below the recommended value a warning is raised.

Change-Id: Ibc43ed2981d6fe4db2f5195a516e1a819c14a6f8